src/main/java/com/typee/logic/interactive/parser/state/undomachine/UndoState.java
package com.typee.logic.interactive.parser.state.undomachine;

import com.typee.logic.commands.Command;
import com.typee.logic.commands.UndoCommand;
import com.typee.logic.commands.exceptions.CommandException;
import com.typee.logic.interactive.parser.ArgumentMultimap;
import com.typee.logic.interactive.parser.Prefix;
import com.typee.logic.interactive.parser.state.EndState;
import com.typee.logic.interactive.parser.state.State;
import com.typee.logic.interactive.parser.state.exceptions.StateTransitionException;

/**
 * Represents the final state of the state machine that builds the {@code UndoCommand}.
 */
public class UndoState extends EndState {

    private static final String MESSAGE_CONSTRAINTS = "The latest undoable command will be undone.";

    public UndoState(ArgumentMultimap soFar) {
        super(soFar);
    }

    @Override
    public Command buildCommand() throws CommandException {
        return new UndoCommand();
    }

    @Override
    public State transition(ArgumentMultimap newArgs) throws StateTransitionException {
        throw new StateTransitionException(MESSAGE_END_STATE);
    }

    @Override
    public String getStateConstraints() {
        return MESSAGE_CONSTRAINTS;
    }

    @Override
    public boolean isEndState() {
        return true;
    }

    @Override
    public Prefix getPrefix() {
        return DUMMY_PREFIX;
    }
}
